Fix signed / unsigned comparison.
Apparently under MinGW timeval.tv_sec (of type time_t) is signed, causing a
signed/unsigned mismatch. I think it's safe to cast to unsigned as that diff
should only ever be +ve, unless time goes backwards.

+++ src/observers/observer_derivative.c
@@ -55,7 +55,7 @@ static void observer_derivative_update( observer_base_t *obs, int first_time )
         gettimeofday( &tv_now, NULL );
         timersub( &tv_now, &this->start_time, &tv_diff );
 
-        if( tv_diff.tv_sec >= this->period )
+        if( (unsigned)tv_diff.tv_sec >= this->period )
         {
             this->derivative = (float)(obs->value - this->value_old) / (float)this->period;
